requirements.txt
opencv-python~=4.9.0.80
# opencv-python~=4.1.1
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
ultralytics==8.1.11
numpy==1.24.4
pillow==10.2.0
nanocamera==0.1.5
python-dotenv==1.0.1
